implementing @nnegrey's feedback
